# api/openapi/kumwe-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kumwe CMS API",
    "version": "1.0.0",
    "description": "Authenticated, explicitly site-bound content management, public discovery, and health endpoints for Kumwe CMS 2.0."
  },
...
  "servers": [{"url": "/"}],
  "paths": {
...
    "/api/v1/content": {
      "get": {
        "operationId": "listContent",
        "summary": "List content",
        "security": [{"bearerAuth": [], "siteContext": []}],
        "parameters": [{
          "name": "include_deleted",
          "in": "query",
          "schema": {"type": "string", "enum": ["0", "1"], "default": "0"}
        }],
        "responses": {
          "200": {
            "description": "Content collection",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentCollection"}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      },
      "post": {
        "operationId": "createContent",
        "summary": "Create a draft",
        "security": [{"bearerAuth": [], "siteContext": []}],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentCreate"}}}
        },
        "responses": {
          "201": {"$ref": "#/components/responses/ContentRecord"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/Unprocessable"}
        }
      }
    },
...
    "/api/v1/content-types": {
      "get": {"operationId": "listContentTypes", "summary": "List published content type heads", "security": [{"bearerAuth": [], "siteContext": []}], "responses": {"200": {"description": "Content type collection"}}},
      "post": {"operationId": "createContentType", "summary": "Create a versioned content type", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ContentTypeMutation"}}}}, "responses": {"201": {"description": "Content type created"}, "422": {"$ref": "#/components/responses/Unprocessable"}}}
    },
...
    "/api/v1/workflows": {
      "get": {"operationId": "listWorkflows", "summary": "List published workflow heads", "security": [{"bearerAuth": [], "siteContext": []}], "responses": {"200": {"description": "Workflow collection"}}},
      "post": {"operationId": "createWorkflow", "summary": "Create a versioned workflow", "security": [{"bearerAuth": [], "siteContext": []}], "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkflowMutation"}}}}, "responses": {"201": {"description": "Workflow created"}, "422": {"$ref": "#/components/responses/Unprocessable"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "Kumwe opaque access token"},
      "siteContext": {
        "type": "apiKey",
        "in": "header",
        "name": "Kumwe-Site",
        "description": "Canonical site identifier. It must exactly match the site bound into the bearer token; host and forwarding headers never select a site."
      }
    },
    "parameters": {
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Unique retry key. A completed response is replayed for 24 hours; reuse with a different request is rejected.",
        "schema": {"type": "string", "minLength": 8, "maxLength": 128, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$"}
      },
...
    },
    "responses": {
      "ContentRecord": {
        "description": "Content record",
        "headers": {
          "ETag": {"description": "Strong version tag", "schema": {"type": "string"}},
          "Idempotency-Replayed": {"description": "Present and true when a stored result was replayed", "schema": {"type": "string"}}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Content"}}}
      },
      "BadRequest": {"description": "Malformed request", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}},
      "Unauthorized": {"description": "Valid bearer token required", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}},
      "Forbidden": {"description": "Token lacks the required capability", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}},
...
      "Conflict": {"description": "Operation is already in progress", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}},
...
      "Unprocessable": {"description": "Validation failed or retry key was reused", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}}
    },
    "schemas": {
      "Content": {
        "type": "object",
        "required": ["id", "title", "slug", "data", "status", "version", "content_type_id", "workflow_id", "created_at", "updated_at"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "title": {"type": "string"},
          "slug": {"type": "string", "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"},
          "data": {"type": "object", "additionalProperties": true},
          "status": {"type": "string", "pattern": "^[a-z][a-z0-9_-]{0,39}$"},
          "publish_at": {"type": ["string", "null"], "format": "date-time"},
          "unpublish_at": {"type": ["string", "null"], "format": "date-time"},
          "version": {"type": "integer", "minimum": 1},
          "content_type_id": {"type": "string", "format": "uuid"},
          "workflow_id": {"type": "string", "format": "uuid"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"},
          "deleted_at": {"type": ["string", "null"], "format": "date-time"}
        },
        "additionalProperties": false
      },
      "ContentCollection": {
        "type": "object",
        "required": ["items"],
        "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/Content"}}},
        "additionalProperties": false
      },
      "ContentCreate": {
        "type": "object",
        "required": ["title", "slug"],
        "properties": {
          "title": {"type": "string", "minLength": 1},
          "slug": {"type": "string", "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$"},
          "content_type": {"type": "string", "description": "Content type UUID or handle; defaults to the core page type"},
          "data": {"type": "object", "additionalProperties": true},
          "publish_at": {"type": ["string", "null"], "format": "date-time"},
          "unpublish_at": {"type": ["string", "null"], "format": "date-time"}
        },
        "additionalProperties": false
      },
...
      "ContentTypeMutation": {
        "type": "object",
        "required": ["name", "workflow", "schema"],
        "properties": {
          "handle": {"type": "string", "pattern": "^[a-z][a-z0-9_-]{0,99}$"},
          "name": {"type": "string", "minLength": 1, "maxLength": 255},
          "workflow": {"type": "string"},
          "schema": {"type": "object"},
          "allow_breaking": {"type": "boolean", "default": false}
        },
        "additionalProperties": false
      },
      "WorkflowMutation": {
        "type": "object",
        "required": ["name", "states", "transitions"],
        "properties": {
          "handle": {"type": "string", "pattern": "^[a-z][a-z0-9_-]{0,99}$"},
          "name": {"type": "string", "minLength": 1, "maxLength": 255},
          "states": {"type": "array", "minItems": 1, "items": {"type": "object"}},
          "transitions": {"type": "array", "items": {"type": "object"}},
          "allow_breaking": {"type": "boolean", "default": false}
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail"],
        "properties": {
          "type": {"type": "string", "format": "uri-reference"},
          "title": {"type": "string"},
          "status": {"type": "integer", "minimum": 400, "maximum": 599},
          "detail": {"type": "string"},
          "instance": {"type": "string", "format": "uri-reference"}
        },
        "additionalProperties": true
      }
    }
  }
}